Combined WebsocketConnection and Connection classes (#323)
