+++ python/gink/impl/sync_func.py
@@ -5,5 +5,5 @@ from .builders import SyncMessage
 
 
 class SyncFunc(Protocol):
-    def __call__(self, *, path: Path, permissions: int, misc: Any) -> SyncMessage:
+    def __call__(self, *, path: Path, perms: int, misc: Any) -> SyncMessage:
         """ Generate the greeting on a new connection. """
